pyproject.toml
[tool.poetry]
name = "subscriptions"
version = "0.1.0"
description = "Проект для курса по Python"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.0"
pydantic = "^2.10.3"
aiohttp = "^3.11.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
